Add new option --neoscrypt-byteswap.
This option enables swapping of the byte order in all four-byte groups
before applying the neoscrypt hash.  (Or, more specifically, we encode
the 32-bit integers in work into big-endian bytes explicitly before
performing the hash.)

This is necessary to make the algo compatible with daemons that directly
call neoscrypt.  See also:

https://cryptocurrencytalk.com/topic/14154-phoenixcoin-neoscrypt-mining-pool/?do=findComment&comment=119575